requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
google-cloud-storage==3.7.0
google-cloud-bigquery==3.39.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
python-dotenv==1.2.1
google-auth-oauthlib==1.2.3
requests==2.32.3